% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rectify.R
\name{measure_mm}
\alias{measure_mm}
\title{Convert a pixel measurement on a rectified image to millimetres}
\usage{
measure_mm(rectified, pixels)
}
\arguments{
\item{rectified}{A \code{"scalecard_rectified"} object.}

\item{pixels}{A numeric distance in pixels measured on the rectified image.}
}
\value{
The distance in millimetres.
}
\description{
Convert a pixel measurement on a rectified image to millimetres
}
